Fix Content-Security-Policy header on /embed route (#260)
The `Content-Security-Policy` header on the `/embed` route currently
only returns the `www.`-prefixed versions of our website URLs, which
causes the embed to fail when entering the website without `www.`. This
PR adds both `www.` and non-`www.`-prefixed versions of the URLs to CSP,
so both will work regardless of how the URLs are configured.

Also added Vitest in order to test this new function.